lax_numpy: refactor set operations into separate private submodule
